partially restoring support for local filesystem captions data

+++ app/src/funcs.php
@@ -1599,9 +1599,8 @@ function captions_data() : PharData
 
 function remove_captions_cache_file(string $filename) : void
 {
-	/*
 	unlink(__DIR__ . '/../captions/' . $filename);
-	*/
+
 	$captions_data = captions_data();
 
 	unset($captions_data[$filename]);
@@ -1609,9 +1608,11 @@ function remove_captions_cache_file(string $filename) : void
 
 function captions_add_from_string(string $filename, string $contents) : void
 {
-	/*
+	if ( ! is_file(__DIR__ . '/../captions/' . $filename)) {
 	file_put_contents(__DIR__ . '/../captions/' . $filename, $contents);
-	*/
+
+		return;
+	}
 	$captions_data = captions_data();
 
 	$captions_data->addFromString($filename, $contents);
@@ -1619,9 +1620,10 @@ function captions_add_from_string(string $filename, string $contents) : void
 
 function captions_get_content(string $filename) : string
 {
-	/*
+	if (is_file(__DIR__ . '/../captions/' . $filename)) {
 	return file_get_contents(__DIR__ . '/../captions/' . $filename);
-	*/
+	}
+
 	$captions_data = captions_data();
 
 	return file_get_contents($captions_data[$filename]->getPathname());
@@ -1629,9 +1631,9 @@ function captions_get_content(string $filename) : string
 
 function captions_content_exists(string $filename) : bool
 {
-	/*
+	if (is_file(__DIR__ . '/../captions/' . $filename)) {
 	return is_file(__DIR__ . '/../captions/' . $filename);
-	*/
+	}
 	$captions_data = captions_data();
 
 	return isset($captions_data[$filename]);
